# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: US Patient App API
  version: 0.1.0
  description: |
    Source-traceable patient-controlled organization API. The service stores
    document metadata, immutable versions, reviewable facts, visit packs,
    tasks, expiring share versions, upload jobs and audit events.
    It does not diagnose, recommend treatment, interpret images, or write to an EHR.
servers:
  - url: https://api.example.invalid
    description: Placeholder; deployment owns the real hostname.
...
security:
  - bearerAuth: [patient]
paths:
  /v1/documents:
    post:
      operationId: createDocument
      tags: [documents]
      security: [{bearerAuth: ['documents:write']}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DocumentCreate'}
      responses:
        '201':
          description: Document metadata created.
          content: {application/json: {schema: {$ref: '#/components/schemas/Document'}}}
        '409': {$ref: '#/components/responses/IdempotencyConflict'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    get:
      operationId: listDocuments
      tags: [documents]
      security: [{bearerAuth: ['documents:read']}]
      responses:
        '200':
          description: Documents visible to the authenticated principal.
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Document'}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The gateway authenticates the principal and maps token claims to scopes.
        The service never logs bearer tokens or document contents.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 128}
...
  responses:
    Unauthorized: {description: Missing or invalid authentication.}
    Forbidden: {description: Authenticated principal lacks scope or ownership.}
...
    IdempotencyConflict: {description: Idempotency-Key was reused with a different request body.}
  schemas:
    SourceType:
      type: string
      enum: [uploaded_document, user_input, ocr, ehr_import, ai_extraction, translation, other]
...
    DocumentStatus:
      type: string
      enum: [uploaded, processing, ready, failed, deleted]
...
    DocumentCreate:
      type: object
      additionalProperties: false
      required: [filename, media_type, size_bytes, sha256]
      properties:
        filename: {type: string, minLength: 1, maxLength: 255}
        media_type: {type: string, minLength: 1, maxLength: 128}
        size_bytes: {type: integer, minimum: 0}
        sha256: {type: string, pattern: '^[A-Fa-f0-9]{64}$'}
    Document:
      allOf:
        - {$ref: '#/components/schemas/DocumentCreate'}
        - type: object
          required: [id, owner_id, source_type, status, version, created_at, updated_at]
          properties:
            id: {type: string, format: uuid}
            owner_id: {type: string}
            source_type: {$ref: '#/components/schemas/SourceType'}
            status: {$ref: '#/components/schemas/DocumentStatus'}
            version: {type: integer, minimum: 1}
            created_at: {type: string, format: date-time}
            updated_at: {type: string, format: date-time}
            deleted_at: {type: string, format: date-time, nullable: true}
...